Add configuration and marketplace data fetching functionality for GPU Marketplace
