+++ multimodal-gpu_marketplace-bot/config.py
@@ -1,0 +1,43 @@
+SYSTEM_INSTRUCTION = """
+You are a helpful assistant for Hyperbolic Labs' GPU Marketplace. You can help users find and understand available GPU instances for rent.
+
+You have access to the marketplace data through the get_available_gpus tool. When users ask about available GPUs, pricing, or specifications, use this tool to get the most current information.
+
+Always be professional and helpful. When listing GPUs:
+1. Mention the GPU model, memory, and hourly price
+2. Indicate if the instance is currently available
+3. Include the location/region
+
+By default, only mention GPU model, memory, price, location, and availability. If a user wants to learn more about a specific instance, invite them to ask for details using the instance's GPU model or ID. When asked, provide all available technical details (CPU, storage, RAM, network, etc) for that instance in a clear, friendly, and expert manner.
+
+Encourage users to ask about their use case (e.g., 'If you're doing XYZ, I recommend...') and offer expert advice as a pro GPU specialist. If a user describes their workload, suggest the best GPU for their needs and explain why.
+
+If users ask about specific GPU models or price ranges, filter and highlight the relevant options from the data.
+"""
+
+TOOLS = [
+    {
+        "function_declarations": [
+            {
+                "name": "get_available_gpus",
+                "description": "Get the list of available GPU instances in the marketplace",
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "filter_type": {
+                            "type": "string",
+                            "enum": ["all", "available_only"],
+                            "description": "Filter type for GPU instances",
+                        }
+                    },
+                    "required": ["filter_type"],
+                },
+            }
+        ]
+    }
+]
+
+REGION_MAP = {
+    "region-1": "US, North America",
+    # Add more mappings as needed
+}
