+++ multimodal-gpu_marketplace-bot/main.py
@@ -27,110 +27,13 @@ from pipecat.services.gemini_multimodal_live.gemini import (
     GeminiMultimodalLiveLLMService,
 )
 from pipecat.transports.services.daily import DailyParams, DailyTransport
+from marketplace import fetch_marketplace_data
+from config import SYSTEM_INSTRUCTION, TOOLS
 
 load_dotenv(override=True)
 
 logger.remove(0)
 logger.add(sys.stderr, level="DEBUG")
-
-
-async def fetch_marketplace_data(
-    function_name, tool_call_id, args, llm, context, result_callback
-):
-    REGION_MAP = {
-        "region-1": "US, North America",
-        # Add more mappings as needed
-    }
-
-    def format_memory(mb):
-        if mb >= 1024 * 1024:
-            return f"{mb / (1024 * 1024):.2f} TB"
-        elif mb >= 1024:
-            return f"{mb / 1024:.2f} GB"
-        else:
-            return f"{mb} MB"
-
-    async with aiohttp.ClientSession() as session:
-        try:
-            url = "https://api.hyperbolic.xyz/v1/marketplace"
-            headers = {"Content-Type": "application/json"}
-            filters = {} if args["filter_type"] == "all" else {"available": True}
-            data = {"filters": filters}
-
-            async with session.post(url, json=data, headers=headers) as response:
-                if response.status == 200:
-                    marketplace_data = await response.json()
-                    available_instances = [
-                        {
-                            # Main summary fields for default listing
-                            "gpu_model": instance["hardware"]["gpus"][0]["model"],
-                            "gpu_memory": format_memory(instance["hardware"]["gpus"][0]["ram"]),
-                            "price_per_hour": f"${instance['pricing']['price']['amount'] / 100:.2f}",
-                            "location": REGION_MAP.get(instance["location"]["region"], instance["location"]["region"]),
-                            "available": not instance["reserved"] and instance["gpus_reserved"] < instance["gpus_total"],
-                            # # All technical details for deep-dive queries
-                            # "id": instance["id"],
-                            # "status": instance.get("status"),
-                            # "hardware": instance.get("hardware"),
-                            # "instances": instance.get("instances"),
-                            # "network": instance.get("network"),
-                            # "gpus_total": instance.get("gpus_total"),
-                            # "gpus_reserved": instance.get("gpus_reserved"),
-                            # "has_persistent_storage": instance.get("has_persistent_storage"),
-                            # "supplier_id": instance.get("supplier_id"),
-                            # "cluster_name": instance.get("cluster_name"),
-                            # "pricing": instance.get("pricing"),
-                        }
-                        for instance in marketplace_data["instances"]
-                        if "gpus" in instance["hardware"] and instance["hardware"]["gpus"]
-                    ]
-                    await result_callback({"instances": available_instances})
-                else:
-                    await result_callback(
-                        {"error": f"API request failed with status {response.status}"}
-                    )
-        except Exception as e:
-            await result_callback({"error": str(e)})
-
-
-tools = [
-    {
-        "function_declarations": [
-            {
-                "name": "get_available_gpus",
-                "description": "Get the list of available GPU instances in the marketplace",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "filter_type": {
-                            "type": "string",
-                            "enum": ["all", "available_only"],
-                            "description": "Filter type for GPU instances",
-                        }
-                    },
-                    "required": ["filter_type"],
-                },
-            }
-        ]
-    }
-]
-
-system_instruction = """
-You are a helpful assistant for Hyperbolic Labs' GPU Marketplace. You can help users find and understand available GPU instances for rent.
-
-You have access to the marketplace data through the get_available_gpus tool. When users ask about available GPUs, pricing, or specifications, use this tool to get the most current information.
-
-Always be professional and helpful. When listing GPUs:
-1. Mention the GPU model, memory, and hourly price
-2. Indicate if the instance is currently available
-3. Include the location/region
-
-By default, only mention GPU model, memory, price, location, and availability. If a user wants to learn more about a specific instance, invite them to ask for details using the instance's GPU model or ID. When asked, provide all available technical details (CPU, storage, RAM, network, etc) for that instance in a clear, friendly, and expert manner.
-
-Encourage users to ask about their use case (e.g., 'If you're doing XYZ, I recommend...') and offer expert advice as a pro GPU specialist. If a user describes their workload, suggest the best GPU for their needs and explain why.
-
-If users ask about specific GPU models or price ranges, filter and highlight the relevant options from the data.
-"""
 
 
 async def main():
@@ -155,8 +58,8 @@ async def main():
 
         llm = GeminiMultimodalLiveLLMService(
             api_key=os.getenv("GOOGLE_API_KEY"),
-            system_instruction=system_instruction,
-            tools=tools,
+            system_instruction=SYSTEM_INSTRUCTION,
+            tools=TOOLS,
         )
 
         llm.register_function("get_available_gpus", fetch_marketplace_data)
